Added OOP & removed lag using after()
Added a class to represent counter and handle it's logic. Also handled the counter instance limit. Used after() to remove lag that came with using window.update(). Ran into GUI freezing when I tried it before using a class. Might still be able to use a stop button. Should still add a function to destroy the counter and recreate it when the start button is clicked if the times have been changed.